refactor(permissions): extract assertCanPostInGroup into group-permissions
